refactor: showState 상태 관련 로직 훅으로 분리
